deploy_files/requirements.txt
# Core dependencies
pandas==2.0.2
numpy==1.24.3
streamlit==1.31.0
python-dotenv==1.0.0

# Web dependencies
gunicorn==21.2.0
protobuf>=3.20.0,<3.21.0

# For dashboard
plotly==5.14.1

# Web3 and blockchain dependencies
web3>=6.0.0
eth-account>=0.8.0
etherscan-python>=2.1.0

# Browser automation (CPU-only version)
browser-use>=0.1.40
playwright>=1.38.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0

# Database
SQLAlchemy>=2.0.0

# Deployment-specific
requests>=2.30.0
aiohttp>=3.8.5
pydantic>=2.0.0
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0 